+++ base/linalg/lapack.jl
@@ -639,8 +639,8 @@ for (gesvx, elty) in
               &fact, &trans, &n, &nrhs, A, &lda, AF, &ldaf, ipiv, &equed, R, C, B,
               &ldb, X, &n, rcond, ferr, berr, work, iwork, info)
             @lapackerror
-            if info[1] == n+1 warn("Matrix is singular to working precision.")
-            else @assertnonsingular
+            if info[1] == n+1 && isinteractive()
+                warn("Matrix is singular to working precision.")
             end
             #WORK(1) contains the reciprocal pivot growth factor norm(A)/norm(U)
             X, equed, R, C, B, rcond[1], ferr, berr, work[1]
@@ -694,8 +694,9 @@ for (gesvx, elty, relty) in
               &fact, &trans, &n, &nrhs, A, &lda, AF, &ldaf, ipiv, &equed, R, C, B,
               &ldb, X, &n, rcond, ferr, berr, work, rwork, info)
             @lapackerror
-            if info[1] == n+1 warn("Matrix is singular to working precision.")
-            else @assertnonsingular end
+            if info[1] == n+1 && isinteractive()
+                warn("Matrix is singular to working precision.")
+            end
             #RWORK(1) contains the reciprocal pivot growth factor norm(A)/norm(U)
             X, equed, R, C, B, rcond[1], ferr, berr, rwork[1]
         end
@@ -1745,7 +1746,7 @@ for (posv, potrf, potri, potrs, pstrf, elty, rtyp) in
             @assertargsok
             #info[1]>0 means the leading minor of order info[i] is not positive definite
             #ordinarily, throw Exception here, but return error code here
-            #this simplifies isposdef! and factorize!
+            #this simplifies isposdef! and factorize
             return A, info[1]
         end
         #       SUBROUTINE DPOTRI( UPLO, N, A, LDA, INFO )
@@ -3181,7 +3182,7 @@ for (bdsdc, elty) in
             if compq == 'N'
                 lwork = 6n
             elseif compq == 'P'
-                warn("COMPQ='P' is not tested")
+                isinteractive() && warn("COMPQ='P' is not tested")
                 #TODO turn this into an actual LAPACK call
                 #smlsiz=ilaenv(9, $elty==:Float64 ? 'dbdsqr' : 'sbdsqr', string(uplo, compq), n,n,n,n)
                 smlsiz=100 #For now, completely overkill