+++ base/string.jl
@@ -19,11 +19,12 @@ next(s::AbstractString, i::Integer) = next(s,int(i))
 function print_to_string(xs...)
     # specialized for performance reasons
     s = IOBuffer(Array(UInt8,isa(xs[1],AbstractString) ? endof(xs[1]) : 0), true, true)
-    truncate(s,0)
     for x in xs
         print(s, x)
     end
-    takebuf_string(s)
+    d = s.data
+    resize!(d,s.size)
+    bytestring(d)
 end
 
 string() = ""
@@ -586,7 +587,6 @@ SubString(s::AbstractString, i::Integer) = SubString(s, i, endof(s))
 
 write{T<:ByteString}(to::IOBuffer, s::SubString{T}) =
     s.endof==0 ? 0 : write_sub(to, s.string.data, s.offset+1, next(s,s.endof)[2]-1)
-print(io::IOBuffer, s::SubString) = write(io, s)
 
 sizeof(s::SubString{ASCIIString}) = s.endof
 sizeof(s::SubString{UTF8String}) = s.endof == 0 ? 0 : next(s,s.endof)[2]-1
@@ -799,7 +799,6 @@ end
 
 endof(s::RopeString) = s.endof
 length(s::RopeString) = length(s.head) + length(s.tail)
-print(io::IO, s::RopeString) = print(io, s.head, s.tail)
 write(io::IO, s::RopeString) = (write(io, s.head); write(io, s.tail))
 sizeof(s::RopeString) = sizeof(s.head) + sizeof(s.tail)