+++ NEWS.md
@@ -76,6 +76,10 @@ Compiler improvements
 Library improvements
 --------------------
 
+  * The `LinAlg` module is now exported.
+
+  * `sparse(A)` now takes any `AbstractMatrix` A as an argument. ([#10031])
+
   * Factorization api is now type-stable, functions dispatch on `Val{false}` or `Val{true}` instead of a boolean value ([#9575]).
 
   * `convert` now checks for overflow when truncating integers or converting between
@@ -170,7 +174,7 @@ Library improvements
 Deprecated or removed
 ---------------------
 
-  * Woodbury special matrix type has been removed from LinAlg ([#10023]).
+  * Woodbury special matrix type has been removed from LinAlg ([#10024]).
 
   * `median` and `median!` no longer accept a `checknan` keyword argument ([#8605]).